// src/types/about.ts
export type AboutFeature = {
  icon: string;
  label: string;
};

export type AboutStat = {
  value: string;
  label: string;
};

export type AboutLogo = {
  src: string;
  alt: string;
};

export type AboutTeamMember = {
  name: string;
  role: string;
  bio: string;
  image: string;
  linkedin?: string;
};

export type AboutFounder = {
  name: string;
  role: string;
  image: string;
  linkedin?: string;
};